+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Upcoming
 Added
 =====
 
+- Add kernel restart prompting to console
 - Add setting to determine backup-on-save behaviour
 - Add basic local kernel which runs in the local interpretor